practice/neural_networks/bccd/requirements.txt
numpy
pandas
Pillow
PyYAML
scikit-learn
torch
torchvision
tqdm
ultralytics
